To Do List - FInal Group Project
Project Requirement:

a. Menu-Driven in the main() - Once the program is executed, display a welcome message to users. Then, using a while loop, display all possible menu options to the users. Once an option is selected, print out the name of the menu. Keep presenting the menu options to the users until an ‘exit condition’ is entered. 

b. Class definition – Define a class for your ‘element’ in a list for your program. Complete all the member variables and member functions for
your class. 

c. Your code should be well structured and easy to read. This includes print NewLine and main function